+++ st2actions/tests/unit/test_remote_runners.py
@@ -20,18 +20,22 @@ tests_config.parse_args()
 import mock
 from unittest2 import TestCase
 
-from st2actions.runners.fabricrunner import get_runner
-from st2actions.runners.fabricrunner import FabricRunner
+from st2actions.runners.fabric_runner import BaseFabricRunner
 from st2common.constants.action import LIVEACTION_STATUS_SUCCEEDED, LIVEACTION_STATUS_FAILED
 from st2common.models.system.action import RemoteScriptAction
 from st2common.models.system.action import FabricRemoteScriptAction
+
+
+class FabricRunner(BaseFabricRunner):
+    def run(self):
+        pass
 
 
 class FabricRunnerTestCase(TestCase):
     def test_get_env_vars(self):
         env_vars = {'key1': 'val1', 'key2': 'val2'}
 
-        runner = get_runner()
+        runner = FabricRunner('id')
         runner.runner_parameters = {'hosts': 'localhost', 'env': env_vars}
         # This is awful, context is just set at some point, no idea when and
         # where MOVE IT TO CONSTRUCTOR!11