+++ st2client/st2client/commands/action.py
@@ -281,6 +281,29 @@ class ActionRunCommandMixin(object):
             # No child error, there might be a global error, include result in the output
             options['attributes'].append('result')
 
+        # On failure we also want to include error message and traceback at the top level
+        if instance.status == 'failed':
+            status_index = options['attributes'].index('status')
+            tasks = instance.result.get('tasks', [])
+
+            top_level_error, top_level_traceback = self._get_top_level_error(live_action=instance)
+            task_error, task_traceback = self._get_task_error(task=tasks[-1])
+
+            if top_level_error:
+                # Top-level error
+                instance.error = top_level_error
+                instance.traceback = top_level_traceback
+                options['attributes'].insert(status_index + 1, 'error')
+                options['attributes'].insert(status_index + 2, 'traceback')
+            elif task_error:
+                # Task error
+                instance.error = task_error
+                instance.traceback = task_traceback
+                instance.failed_on = tasks[-1].get('name', 'unknown')
+                options['attributes'].insert(status_index + 1, 'error')
+                options['attributes'].insert(status_index + 2, 'traceback')
+                options['attributes'].insert(status_index + 3, 'failed_on')
+
         # print root task
         self.print_output(instance, formatter, **options)
 
@@ -315,6 +338,34 @@ class ActionRunCommandMixin(object):
                 execution.result = self._format_error_result(execution.result)
 
         return execution
+
+    def _get_top_level_error(self, live_action):
+        """
+        Retrieve a top level workflow error.
+
+        :return: (error, traceback)
+        """
+        error = live_action.result.get('error', None)
+        traceback = live_action.result.get('traceback', None)
+
+        return error, traceback
+
+    def _get_task_error(self, task):
+        """
+        Retrieve error message from the provided task.
+
+        :return: (error, traceback)
+        """
+        if not task:
+            return None, None
+
+        result = task['result']
+        stderr = result.get('stderr', None)
+        error = result.get('error', None)
+        traceback = result.get('traceback', None)
+        error = error if error else stderr
+
+        return error, traceback
 
     def _is_error_result(self, result):
         if not isinstance(result, dict):
@@ -921,7 +972,6 @@ class ActionExecutionReRunCommand(ActionRunCommandMixin, resource.ResourceComman
         runner_mgr = self.app.client.managers['RunnerType']
         action_exec_mgr = self.app.client.managers['LiveAction']
 
-        # TODO use action.ref when this attribute is added
         action_ref = existing_execution.action['ref']
         action = action_mgr.get_by_ref_or_id(action_ref)
         runner = runner_mgr.get_by_name(action.runner_type)
@@ -929,16 +979,8 @@ class ActionExecutionReRunCommand(ActionRunCommandMixin, resource.ResourceComman
         action_parameters = self._get_action_parameters_from_args(action=action, runner=runner,
                                                                   args=args)
 
-        # Create new execution object
-        new_execution = models.LiveAction()
-        new_execution.action = action_ref
-        new_execution.parameters = getattr(existing_execution, 'parameters', {})
-
-        # If user provides parameters merge and override with the ones from the
-        # existing execution
-        new_execution.parameters.update(action_parameters)
-
-        execution = action_exec_mgr.create(new_execution, **kwargs)
+        execution = action_exec_mgr.re_run(execution_id=args.id, parameters=action_parameters,
+                                           **kwargs)
         execution = self._get_execution_result(execution=execution,
                                                action_exec_mgr=action_exec_mgr,
                                                args=args, **kwargs)