+++ st2client/st2client/models/__init__.py
@@ -17,5 +17,6 @@ from st2client.models.core import *         # noqa
 from st2client.models.access import *       # noqa
 from st2client.models.action import *       # noqa
 from st2client.models.datastore import *    # noqa
+from st2client.models.policy import *       # noqa
 from st2client.models.reactor import *      # noqa
 from st2client.models.webhook import *      # noqa